test reordering
